KDE4 no more progress bar flickering
This will always redraw the progress bar over the last one.
It stops the flickering at least in KDE4 from re-painting the
background, but obvously will break any transparent or alaiasing
effect, if it was used.

There is actually a mbProgressNeedsErase in the global settings,
so actually use it when setting a new progress value.

Change-Id: Ic62784da4f6d08bf0e03a9e04848a5cc687ee236